requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.11
langchain-google-genai==0.0.4
langchain-core==0.1.12
python-dotenv==1.0.0
faiss-cpu==1.7.4
unstructured==0.10.30
pypdf==3.17.4
Jinja2==3.1.2